api gateway and lambda integration done

+++ infra/trapply.sh
@@ -1,4 +1,6 @@
 source /usr/local/bin/awsha
+if [ ! -f appstack.json ]; then
 echo "{}" > appstack.json
+fi
 terraform apply plan.out
